third_party/blink/web_tests/crypto/subtle/rsassa-pkcs1-v1_5/import-jwk-expected.txt

Test importing an RSA key for RSASSA-PKCS1-v1_5.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Importing a public key...
PASS publicKey.toString() is '[object CryptoKey]'
PASS publicKey.type is 'public'
PASS publicKey.usages is ['verify']
PASS publicKey.algorithm.name is 'RSASSA-PKCS1-v1_5'
PASS publicKey.algorithm.modulusLength is 2048
PASS bytesToHexString(publicKey.algorithm.publicExponent) is '010001'
PASS publicKey.algorithm.hash.name is 'SHA-256'

Importing a private key...
PASS privateKey.toString() is '[object CryptoKey]'
PASS privateKey.type is 'private'
PASS privateKey.usages is ['sign']
PASS privateKey.algorithm.name is 'RSASSA-PKCS1-v1_5'
PASS privateKey.algorithm.modulusLength is 2048
PASS bytesToHexString(privateKey.algorithm.publicExponent) is '010001'
PASS privateKey.algorithm.hash.name is 'SHA-256'
PASS successfullyParsed is true

TEST COMPLETE
